Refactor GearChangeFrame into GearChangePanel.
